Fix -Werror=sign-compare

+++ librandom/getrandom.c
@@ -11,7 +11,7 @@
 /// \return 0 on success or 1 in case of an error
 uint8_t crypto_api_osrandom_native(uint8_t* buf, size_t len) {
 	// Get the random bytes
-	if (getrandom(buf, len, 0) != len) {
+	if (getrandom(buf, len, 0) != (ssize_t)len) {
 		return 1;
 	}
